LIBFCREPO-894. Retrieve "umd-camel-processors" dependency from Nexus
Added a "activemq/maven-settings.xml" file containing the Maven
"settings.xml" configuration needed for accessing the UMD Maven Nexus.

In the "Dockerfile", modified the "dependencies" Docker stage by adding
a "maven-settings.xml" file with the configuration necessary to access
the UMD Maven Nexus. This eliminates the need to copy over a shaded
"uber-jar" in "umd-camel-processors" -- the necessary jar can now be
retrieved from the Nexus.

Modified the "pom.xml" to add the "umd-camel-processors" jar as a
dependency.

https://issues.umd.edu/browse/LIBFCREPO-894