LIBFCREPO-905. Made "index-helper.js" default updateRequestProcessorChain
Modified the "solrconfig.xml" to make "index-helper.js" script the
default updateRequestProcessorChain for the Solr configuration.

This is needed, as otherwise the "index-helper.js" script will not be
called when adding/updating Solr entries, which is needed because the
script adjusts/populates some fields.

Reverted the changes made to the "schema.xml" as they are not needed
due when the "index-helper.js" script is working correctly.

https://issues.umd.edu/browse/LIBFCREPO-905