LIBFCREPO-947. Remove plastrond Docker configuration
Removed plastrond container from the "umd-fcrepo" stack because running
Plastron in the Docker stack does not currently work in the local
development environment.

1) When running in a local development environment, the daemon needs to
access "import" and "export" files in directories on the workstation,
which is not currently possible from inside a Docker container in a
stack.

2) When running in a local development environment, the fcrepo webapp i
 accessed at "http://localhost:8080" from the workstation. When running
in a Docker container, Plastron cannot access the fcrepo webapp, because
the "http://localhost:8080" URL is assumed by the container to be
running in the container.

While it is likely both of the above issues are fixable (and it is
desirable to have the Plastron daemon run as part of the stack), at this
particular moment it is not possible to devote the time/resources to
finding solutions.

Plastron will need to be run as a separate application when used with
the local development environment.

https://issues.umd.edu/browse/LIBFCREPO-947